requirements.txt
sse-starlette==2.0.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
streamlit==1.37.1
pandas==2.2.2
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.8
asgi-lifespan==2.1.0
typing-extensions==4.12.2
